Modification de la BDD pour autoriser l'achat par un meme personne de plusieurs Billets identique (problème de cle primaire)

+++ src/Modele/Achat.java
@@ -123,8 +123,8 @@ public class Achat {
 	public int getPrixUnitaire() {
 		return (Integer) map.get("prix_unitaire");
 	}
-	public int getPrixTotal() {
-		return (Integer) map.get("prix_total");
+	public double getPrixTotal() {
+		return (Double) map.get("prix_total");
 	}
 	public Personne getPersonne() {
 		return personne;
